Make handwritten vertex decoders work with non-compiled vertex decoding
